server/requirements.txt
Django==4.2.23
djangorestframework
drf-spectacular
djangorestframework-simplejwt
celery
redis
django-cors-headers
openai
gunicorn
yfinance
google-api-python-client
google-auth-httplib2
google-auth-oauthlib